requirements.txt
numpy
hydra-core
Pillow
tqdm
h5py
opencv-python
open3d